:truck: Move Picasa store to misc/deprecated
And also:
  - :fire: remove related picasa documentation (rst files)
  - :wrench: Adapt config files to the removal of Picasa
