--- docs/source/backends/picasa.rst
@@ -1,7 +1,0 @@
-
-.. _coherence.backends.picasa:
-
-coherence.backends.picasa\_storage (module)
--------------------------------------------
-
-.. automodule:: coherence.backends.picasa_storage
