+++ pyproject.toml
@@ -114,7 +114,6 @@ ITVStore = 'coherence.backends.itv_storage:ITVStore'
 LastFMStore = 'coherence.backends.lastfm_storage:LastFMStore'
 LolcatsStore = 'coherence.backends.lolcats_storage:LolcatsStore'
 MediaStore = 'coherence.backends.mediadb_storage:MediaStore'
-PicasaStore = 'coherence.backends.picasa_storage:PicasaStore'
 PlaylistStore = 'coherence.backends.playlist_storage:PlaylistStore'
 RadiotimeStore = 'coherence.backends.radiotime_storage:RadiotimeStore'
 SWR3Store = 'coherence.backends.swr3_storage:SWR3Store'
